fix(auth): implement RBAC, resolve edge runtime crypto issues, and migrate middleware to proxy
