+++ src/auth.ts
@@ -1,5 +1,6 @@
 import NextAuth, { type DefaultSession } from "next-auth";
 import { PrismaAdapter } from "@auth/prisma-adapter";
+import Credentials from "next-auth/providers/credentials";
 import { prisma } from "@/lib/core/prisma";
 import authConfig from "./auth.config";
 import { extractAccountName } from "@/lib/utils/utils";
@@ -25,6 +26,40 @@ export const { handlers, auth, signIn, signOut } = NextAuth({
     signIn: "/login",
   },
   ...authConfig,
+  providers: [
+    ...authConfig.providers,
+    ...(process.env.NEXT_PUBLIC_E2E === 'true' && process.env.NODE_ENV === 'development' ? [
+      Credentials({
+        name: "E2E Credentials",
+        credentials: {
+          email: { label: "Email", type: "email" },
+          password: { label: "Password", type: "password" }
+        },
+        async authorize(credentials) {
+          const expectedPassword = process.env.E2E_TEST_PASSWORD || 'social-studio-e2e-secret';
+          
+          if (
+            credentials?.email === "tester@socialstudio.ai" && 
+            credentials?.password === expectedPassword
+          ) {
+            const user = await prisma.user.findFirst({
+              where: { email: "tester@socialstudio.ai" }
+            });
+
+            if (user) {
+              return {
+                id: user.id,
+                name: user.name || "E2E Tester",
+                email: user.email,
+                role: user.role
+              };
+            }
+          }
+          return null;
+        }
+      })
+    ] : []),
+  ],
   callbacks: {
     ...authConfig.callbacks,
     async session({ session, token }) {
