+++ src/lib/core/encryption.ts
@@ -1,72 +1,46 @@
-import crypto from 'crypto';
-
 /**
  * ENCRYPTION UTILITY
- * Provides AES-256-GCM encryption/decryption for sensitive database fields.
- * Uses AUTH_SECRET as the master key.
  */
+
+// Use dynamic require to avoid Edge Runtime error when this file is imported
+const crypto = typeof window === 'undefined' ? require('crypto') : null;
 
 const ALGORITHM = 'aes-256-gcm';
 const IV_LENGTH = 12;
-const TAG_LENGTH = 16;
 
-// Derive a 32-byte key from the AUTH_SECRET
 function getEncryptionKey() {
-  const secret = process.env.AUTH_SECRET;
-  if (!secret) {
-    if (process.env.NODE_ENV === 'production') {
-      throw new Error('AUTH_SECRET is not set in production!');
-    }
-    // Fallback for local development if not set
-    return crypto.createHash('sha256').update('fallback-local-secret').digest();
-  }
+  const secret = process.env.AUTH_SECRET || 'fallback-local-secret';
   return crypto.createHash('sha256').update(secret).digest();
 }
 
 /**
- * Encrypts a string into a format suitable for storage.
- * Output format: [iv_hex]:[auth_tag_hex]:[encrypted_data_hex]
+ * Synchronous encryption/decryption (Node.js only, used by Prisma)
  */
 export function encrypt(text: string): string {
-  if (!text) return text;
-  
+  if (!text || !crypto) return text;
   const iv = crypto.randomBytes(IV_LENGTH);
   const key = getEncryptionKey();
   const cipher = crypto.createCipheriv(ALGORITHM, key, iv);
-  
   let encrypted = cipher.update(text, 'utf8', 'hex');
   encrypted += cipher.final('hex');
-  
   const authTag = cipher.getAuthTag().toString('hex');
-  
   return `${iv.toString('hex')}:${authTag}:${encrypted}`;
 }
 
-/**
- * Decrypts a previously encrypted string.
- */
 export function decrypt(encryptedText: string): string {
-  if (!encryptedText) return encryptedText;
-  
-  // If the text doesn't match our format, return as-is (graceful migration/fallback)
-  if (!encryptedText.includes(':')) return encryptedText;
-
+  if (!encryptedText || !encryptedText.includes(':') || !crypto) return encryptedText;
   try {
-    const [ivHex, authTagHex, encryptedData] = encryptedText.split(':');
-    
+    const [ivHex, tagHex, dataHex] = encryptedText.split(':');
     const iv = Buffer.from(ivHex, 'hex');
-    const authTag = Buffer.from(authTagHex, 'hex');
+    const tag = Buffer.from(tagHex, 'hex');
     const key = getEncryptionKey();
-    
     const decipher = crypto.createDecipheriv(ALGORITHM, key, iv);
-    decipher.setAuthTag(authTag);
-    
-    let decrypted = decipher.update(encryptedData, 'hex', 'utf8');
+    decipher.setAuthTag(tag);
+    let decrypted = decipher.update(dataHex, 'hex', 'utf8');
     decrypted += decipher.final('utf8');
-    
     return decrypted;
   } catch (error) {
-    console.error('Decryption failed. Returning raw value.', error);
+    console.error('Decryption failed.', error);
     return encryptedText;
   }
 }
